refactor(git): git/mod.rs 拆五个子模块
- 2720→61 行（mod 声明+重导出）：remote_forward 484（含 cfg(test) 矩阵）/diff_collect 832/push 93/pr_workflow_helpers 527/tests 874
- 关键耦合处理：commands.rs 的 use super::* glob 消费面经 mod.rs pub(crate) 重导出穷尽
- TokenIsolatedCommandOutput 等 struct 字段升 pub(crate)
- 给 git/commands.rs(2343) 立 2000 行红线在 policy（下一步）
- baseline 重生成 + 新文件登记豁免
验证：cargo check 零 error；cargo test --lib git 98 passed；新文件 rustfmt 过；gate 绿
