+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T19:00:31.216Z",
+  "generatedAt": "2026-08-30T19:15:50.955Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -381,16 +381,6 @@
       "priority": "P1",
       "warnThreshold": 2400,
       "failThreshold": 2800,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src-tauri/src/git/mod.rs",
-      "lines": 2720,
-      "type": "rust",
-      "policyId": "bridge-runtime-critical",
-      "priority": "P0",
-      "warnThreshold": 2200,
-      "failThreshold": 2600,
       "thresholdSource": "policy"
     },
     {
