+++ src-tauri/src/git/mod.rs
@@ -41,2680 +41,18 @@ mod pull_request_content;
 mod range_gate;
 mod validation;
 
-#[cfg(test)]
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct GitRemoteForwardingEntry {
-    pub(crate) method: &'static str,
-    pub(crate) category: &'static str,
-    pub(crate) desktop_module: &'static str,
-    pub(crate) daemon_dispatch: &'static str,
-    pub(crate) forwarding: &'static str,
-    pub(crate) coverage: &'static str,
-}
-
-#[cfg(test)]
-pub(crate) const GIT_REMOTE_FORWARDING_MATRIX: &[GitRemoteForwardingEntry] = &[
-    GitRemoteForwardingEntry {
-        method: "get_git_status",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_status",
-        forwarding: "implemented",
-        coverage: "representative-read",
-    },
-    GitRemoteForwardingEntry {
-        method: "list_git_roots",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "list_git_roots",
-        forwarding: "implemented",
-        coverage: "issue-633-root-scan",
-    },
-    GitRemoteForwardingEntry {
-        method: "list_git_repository_summaries",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "list_git_repository_summaries",
-        forwarding: "implemented",
-        coverage: "multi-repository-summary",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_diffs",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_diffs",
-        forwarding: "implemented",
-        coverage: "read-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_file_full_diff",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_file_full_diff",
-        forwarding: "implemented",
-        coverage: "read-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_file_blame",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_file_blame",
-        forwarding: "implemented",
-        coverage: "file-view-blame",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_remote",
-        category: "read",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_remote",
-        forwarding: "implemented",
-        coverage: "read-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_log",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_log",
-        forwarding: "implemented",
-        coverage: "history-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_commit_history",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_commit_history",
-        forwarding: "implemented",
-        coverage: "history-filter-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "resolve_git_commit_ref",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "resolve_git_commit_ref",
-        forwarding: "implemented",
-        coverage: "history-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_commit_details",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_commit_details",
-        forwarding: "implemented",
-        coverage: "history-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_commit_diff",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_commit_diff",
-        forwarding: "implemented",
-        coverage: "history-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_push_preview",
-        category: "history",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_push_preview",
-        forwarding: "implemented",
-        coverage: "push-preview-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "stage_git_file",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "stage_git_file",
-        forwarding: "implemented",
-        coverage: "representative-write",
-    },
-    GitRemoteForwardingEntry {
-        method: "stage_git_all",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "stage_git_all",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "unstage_git_file",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "unstage_git_file",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "unstage_git_all",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "unstage_git_all",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "unstage_git_paths",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "unstage_git_paths",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "revert_git_file",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "revert_git_file",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "revert_git_paths",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "revert_git_paths",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "revert_git_all",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "revert_git_all",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "commit_git",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "commit_git",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "push_git",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "push_git",
-        forwarding: "implemented",
-        coverage: "push-params-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "pull_git",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "pull_git",
-        forwarding: "implemented",
-        coverage: "pull-params-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "sync_git",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "sync_git",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "git_pull",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "git_pull",
-        forwarding: "implemented",
-        coverage: "legacy-toolbar-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "git_push",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "git_push",
-        forwarding: "implemented",
-        coverage: "legacy-toolbar-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "git_sync",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "git_sync",
-        forwarding: "implemented",
-        coverage: "legacy-toolbar-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "git_fetch",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "git_fetch",
-        forwarding: "implemented",
-        coverage: "fetch-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "update_git_branch",
-        category: "write",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "update_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-update-local-regression",
-    },
-    GitRemoteForwardingEntry {
-        method: "cherry_pick_commit",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "cherry_pick_commit",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "revert_commit",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "revert_commit",
-        forwarding: "implemented",
-        coverage: "write-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "reset_git_commit",
-        category: "write",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "reset_git_commit",
-        forwarding: "implemented",
-        coverage: "reset-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "list_git_branches",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "list_git_branches",
-        forwarding: "implemented",
-        coverage: "branch-list-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "checkout_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "checkout_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "create_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "create_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "create_git_branch_from_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "create_git_branch_from_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "create_git_branch_from_commit",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "create_git_branch_from_commit",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "delete_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "delete_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "rename_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "rename_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "merge_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "merge_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "rebase_git_branch",
-        category: "branch",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "rebase_git_branch",
-        forwarding: "implemented",
-        coverage: "branch-mutation-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_branch_compare_commits",
-        category: "worktree",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "get_git_branch_compare_commits",
-        forwarding: "implemented",
-        coverage: "compare-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_branch_diff_between_branches",
-        category: "worktree",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "get_git_branch_diff_between_branches",
-        forwarding: "implemented",
-        coverage: "compare-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_branch_file_diff_between_branches",
-        category: "worktree",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "get_git_branch_file_diff_between_branches",
-        forwarding: "implemented",
-        coverage: "compare-path-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_worktree_diff_against_branch",
-        category: "worktree",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "get_git_worktree_diff_against_branch",
-        forwarding: "implemented",
-        coverage: "worktree-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_worktree_file_diff_against_branch",
-        category: "worktree",
-        desktop_module: "commands_branch.rs",
-        daemon_dispatch: "get_git_worktree_file_diff_against_branch",
-        forwarding: "implemented",
-        coverage: "worktree-path-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_github_issues",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_github_issues",
-        forwarding: "implemented",
-        coverage: "github-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_github_pull_requests",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_github_pull_requests",
-        forwarding: "implemented",
-        coverage: "github-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_github_pull_request_diff",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_github_pull_request_diff",
-        forwarding: "implemented",
-        coverage: "github-pr-number-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_github_pull_request_comments",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_github_pull_request_comments",
-        forwarding: "implemented",
-        coverage: "github-pr-number-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "get_git_pr_workflow_defaults",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "get_git_pr_workflow_defaults",
-        forwarding: "implemented",
-        coverage: "pr-workflow-matrix",
-    },
-    GitRemoteForwardingEntry {
-        method: "create_git_pr_workflow",
-        category: "github",
-        desktop_module: "commands.rs",
-        daemon_dispatch: "create_git_pr_workflow",
-        forwarding: "implemented",
-        coverage: "pr-workflow-params-matrix",
-    },
-];
-async fn should_forward_git_remote(state: &State<'_, AppState>) -> bool {
-    crate::remote_backend::is_remote_mode(&*state).await
-}
-
-async fn forward_git_remote<T: DeserializeOwned>(
-    state: &State<'_, AppState>,
-    app: &AppHandle,
-    method: &str,
-    params: serde_json::Value,
-) -> Result<T, String> {
-    let response = crate::remote_backend::call_remote(&*state, app.clone(), method, params).await?;
-    serde_json::from_value(response).map_err(|error| error.to_string())
-}
-
-async fn forward_git_remote_unit(
-    state: &State<'_, AppState>,
-    app: &AppHandle,
-    method: &str,
-    params: serde_json::Value,
-) -> Result<(), String> {
-    let _: serde_json::Value = forward_git_remote(state, app, method, params).await?;
-    Ok(())
-}
-
-const INDEX_SKIP_WORKTREE_FLAG: u16 = 0x4000;
-const MAX_COMMIT_DIFF_LINES: usize = 10_000;
-const GIT_COMMAND_TIMEOUT_SECS: u64 = 120;
-const GIT_STATUS_DIFF_STATS_FILE_LIMIT: usize = 120;
-const GIT_STATUS_DIFF_STATS_MAX_FILE_BYTES: u64 = 256 * 1024;
-const GIT_DIFF_PREVIEW_MAX_FILES: usize = 200;
-const GIT_DIFF_PREVIEW_MAX_TOTAL_BYTES: usize = 2 * 1024 * 1024;
-const GIT_DIFF_PREVIEW_MAX_BYTES_PER_FILE: usize = 256 * 1024;
-const GIT_DIFF_PREVIEW_MAX_LINES_PER_FILE: usize = 2_500;
-const GIT_DIFF_PREVIEW_SKIP_FILE_SIZE_BYTES: u64 = 1024 * 1024;
-
-fn trim_lowercase(input: Option<String>) -> Option<String> {
-    input
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-        .map(|value| value.to_lowercase())
-}
-
-fn trim_optional(input: Option<String>) -> Option<String> {
-    input
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
-fn is_branch_used_by_worktree_error(raw: &str) -> bool {
-    let message = raw.to_lowercase();
-    message.contains("cannot delete branch") && message.contains("used by worktree at")
-}
-
-fn extract_worktree_path_from_delete_error(raw: &str) -> Option<String> {
-    let marker = "used by worktree at '";
-    let start = raw.find(marker)?;
-    let tail = &raw[start + marker.len()..];
-    let end = tail.find('\'')?;
-    let path = tail[..end].trim();
-    if path.is_empty() {
-        return None;
-    }
-    Some(path.to_string())
-}
-
-fn build_delete_branch_worktree_error(branch_name: &str, raw: &str) -> String {
-    if let Some(path) = extract_worktree_path_from_delete_error(raw) {
-        return format!(
-            "Cannot delete branch '{branch_name}' because it is currently used by worktree at '{path}'. Switch that worktree to another branch or remove that worktree, then retry."
-        );
-    }
-    format!(
-        "Cannot delete branch '{branch_name}' because it is currently used by another worktree. Switch that worktree to another branch or remove that worktree, then retry."
-    )
-}
-
-fn truncate_diff_lines(content: &str, max_lines: usize) -> (String, usize, bool) {
-    if max_lines == 0 {
-        return (String::new(), 0, false);
-    }
-    let mut lines = content.lines();
-    let mut kept = Vec::new();
-    let mut total = 0usize;
-    let mut truncated = false;
-    for line in lines.by_ref() {
-        total += 1;
-        if total <= max_lines {
-            kept.push(line);
-        } else {
-            truncated = true;
-        }
-    }
-    (kept.join("\n"), total, truncated || total > max_lines)
-}
-
-fn normalize_guard_path(path: &str) -> String {
-    path.replace('\\', "/").to_ascii_lowercase()
-}
-
-fn is_heavy_diff_path(path: &str) -> bool {
-    let normalized = normalize_guard_path(path);
-    let segments: Vec<&str> = normalized
-        .split('/')
-        .filter(|segment| !segment.is_empty())
-        .collect();
-    let file_name = segments.last().copied().unwrap_or(normalized.as_str());
-
-    if matches!(
-        file_name,
-        "pnpm-lock.yaml"
-            | "package-lock.json"
-            | "yarn.lock"
-            | "bun.lockb"
-            | "cargo.lock"
-            | "pipfile.lock"
-            | "poetry.lock"
-            | "composer.lock"
-    ) {
-        return true;
-    }
-
-    if file_name.ends_with(".lock")
-        || file_name.ends_with(".min.js")
-        || file_name.ends_with(".bundle.js")
-    {
-        return true;
-    }
-
-    segments.iter().any(|segment| {
-        matches!(
-            *segment,
-            "node_modules"
-                | ".pnpm"
-                | ".pnpm-store"
-                | ".next"
-                | "dist"
-                | "build"
-                | "coverage"
-                | "release-artifacts"
-        )
-    })
-}
-
-fn is_large_worktree_file(repo_root: &Path, path: &str, limit_bytes: u64) -> bool {
-    let candidate = repo_root.join(path);
-    match fs::metadata(candidate) {
-        Ok(metadata) => metadata.is_file() && metadata.len() > limit_bytes,
-        Err(_) => false,
-    }
-}
-
-fn should_skip_diff_stats(repo_root: &Path, path: &str) -> bool {
-    is_heavy_diff_path(path)
-        || is_large_worktree_file(repo_root, path, GIT_STATUS_DIFF_STATS_MAX_FILE_BYTES)
-}
-
-fn utf8_safe_prefix(input: &str, max_bytes: usize) -> &str {
-    if input.len() <= max_bytes {
-        return input;
-    }
-    let mut end = max_bytes;
-    while end > 0 && !input.is_char_boundary(end) {
-        end -= 1;
-    }
-    &input[..end]
-}
-
-fn truncate_diff_preview(content: String, max_lines: usize, max_bytes: usize) -> String {
-    let (mut trimmed, _total_lines, line_truncated) = truncate_diff_lines(&content, max_lines);
-    let mut truncated = line_truncated;
-
-    if trimmed.len() > max_bytes {
-        let safe_prefix = utf8_safe_prefix(&trimmed, max_bytes).to_string();
-        trimmed = safe_prefix;
-        truncated = true;
-    }
-
-    if truncated {
-        trimmed.push_str("\n\n[diff truncated for performance]");
-    }
-
-    trimmed
-}
-
-fn collect_commit_refs_map(repo: &Repository) -> HashMap<Oid, Vec<String>> {
-    let mut map: HashMap<Oid, Vec<String>> = HashMap::new();
-    let references = match repo.references() {
-        Ok(references) => references,
-        Err(_) => return map,
-    };
-    for reference in references.flatten() {
-        let oid = match reference.target() {
-            Some(oid) => oid,
-            None => continue,
-        };
-        let name = reference
-            .shorthand()
-            .or_else(|| reference.name())
-            .unwrap_or("")
-            .to_string();
-        if name.is_empty() {
-            continue;
-        }
-        map.entry(oid).or_default().push(name);
-    }
-    for values in map.values_mut() {
-        values.sort();
-        values.dedup();
-    }
-    map
-}
-
-fn open_repository_at_root(repo_root: &Path) -> Result<Repository, String> {
-    Repository::open_ext(
-        repo_root,
-        git2::RepositoryOpenFlags::NO_SEARCH,
-        std::iter::empty::<&Path>(),
-    )
-    .map_err(|e| e.to_string())
-}
-
-fn paginate_history_commits(
-    commits: Vec<GitHistoryCommit>,
-    offset: usize,
-    limit: usize,
-) -> (Vec<GitHistoryCommit>, usize, bool) {
-    let total = commits.len();
-    let page: Vec<GitHistoryCommit> = commits.into_iter().skip(offset).take(limit).collect();
-    let has_more = offset.saturating_add(page.len()) < total;
-    (page, total, has_more)
-}
-
-fn resolve_ref_to_oid(repo: &Repository, reference: &str) -> Result<Oid, String> {
-    let trimmed = reference.trim();
-    if trimmed.is_empty() {
-        return Err("Branch name cannot be empty.".to_string());
-    }
-    let local_ref = format!("refs/heads/{trimmed}");
-    if let Ok(oid) = repo.refname_to_id(&local_ref) {
-        return Ok(oid);
-    }
-    let remote_ref = format!("refs/remotes/{trimmed}");
-    if let Ok(oid) = repo.refname_to_id(&remote_ref) {
-        return Ok(oid);
-    }
-    repo.revparse_single(trimmed)
-        .map(|object| object.id())
-        .map_err(|_| format!("Branch or ref not found: {trimmed}"))
-}
-
-fn commit_to_history_commit(
-    commit: &git2::Commit<'_>,
-    refs_map: &HashMap<Oid, Vec<String>>,
-) -> GitHistoryCommit {
-    let oid = commit.id();
-    let sha = oid.to_string();
-    let short_sha: String = sha.chars().take(7).collect();
-    let summary = commit.summary().unwrap_or("").to_string();
-    let message = commit.message().unwrap_or("").to_string();
-    let author = commit.author().name().unwrap_or("").to_string();
-    let author_email = commit.author().email().unwrap_or("").to_string();
-    let timestamp = commit.time().seconds();
-    let parents = commit
-        .parents()
-        .map(|parent| parent.id().to_string())
-        .collect();
-    let refs = refs_map.get(&oid).cloned().unwrap_or_default();
-    GitHistoryCommit {
-        sha,
-        short_sha,
-        summary,
-        message,
-        author,
-        author_email,
-        timestamp,
-        parents,
-        refs,
-        file_path: None,
-    }
-}
-
-fn collect_unique_commits(
-    repo: &Repository,
-    include_ref: &str,
-    exclude_ref: &str,
-    refs_map: &HashMap<Oid, Vec<String>>,
-    limit: usize,
-) -> Result<Vec<GitHistoryCommit>, String> {
-    let include_oid = resolve_ref_to_oid(repo, include_ref)?;
-    let exclude_oid = resolve_ref_to_oid(repo, exclude_ref)?;
-    let mut revwalk = repo.revwalk().map_err(|e| e.to_string())?;
-    revwalk
-        .set_sorting(Sort::TOPOLOGICAL | Sort::TIME)
-        .map_err(|e| e.to_string())?;
-    revwalk.push(include_oid).map_err(|e| e.to_string())?;
-    revwalk.hide(exclude_oid).map_err(|e| e.to_string())?;
-
-    let mut commits = Vec::new();
-    for oid_result in revwalk.take(limit) {
-        let oid = oid_result.map_err(|e| e.to_string())?;
-        let commit = repo.find_commit(oid).map_err(|e| e.to_string())?;
-        commits.push(commit_to_history_commit(&commit, refs_map));
-    }
-    Ok(commits)
-}
-
-fn parse_remote_branch(name: &str) -> Option<(String, String)> {
-    let trimmed = name.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    let without_prefix = trimmed
-        .strip_prefix("refs/remotes/")
-        .or_else(|| trimmed.strip_prefix("remotes/"))
-        .unwrap_or(trimmed);
-    let mut parts = without_prefix.splitn(2, '/');
-    let remote = parts.next()?.trim();
-    let branch = parts.next()?.trim();
-    if remote.is_empty() || branch.is_empty() {
-        return None;
-    }
-    Some((remote.to_string(), branch.to_string()))
-}
-
-async fn run_git_command(repo_root: &Path, args: &[&str]) -> Result<(), String> {
-    let git_bin = resolve_git_binary().map_err(|e| format!("Failed to run git: {e}"))?;
-    let mut command = crate::utils::async_command(git_bin);
-    command
-        .args(args)
-        .current_dir(repo_root)
-        .env("PATH", git_env_path())
-        // Force non-interactive git in GUI context so pull/fetch does not hang on hidden prompts.
-        .env("GIT_TERMINAL_PROMPT", "0")
-        .env("GCM_INTERACTIVE", "never")
-        .stdin(Stdio::null())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-
-    let output = match timeout(
-        Duration::from_secs(GIT_COMMAND_TIMEOUT_SECS),
-        command.output(),
-    )
-    .await
-    {
-        Ok(result) => result.map_err(|e| format!("Failed to run git: {e}"))?,
-        Err(_) => {
-            let command_name = args.join(" ");
-            return Err(format!(
-                "Git command timed out after {GIT_COMMAND_TIMEOUT_SECS}s: git {command_name}. Check network/authentication and retry."
-            ));
-        }
-    };
-
-    if output.status.success() {
-        return Ok(());
-    }
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let detail = if stderr.trim().is_empty() {
-        stdout.trim()
-    } else {
-        stderr.trim()
-    };
-    if detail.is_empty() {
-        return Err("Git command failed.".to_string());
-    }
-    Err(detail.to_string())
-}
-
-fn parse_upstream_ref(name: &str) -> Option<(String, String)> {
-    let trimmed = name.strip_prefix("refs/remotes/").unwrap_or(name);
-    let mut parts = trimmed.splitn(2, '/');
-    let remote = parts.next()?;
-    let branch = parts.next()?;
-    if remote.is_empty() || branch.is_empty() {
-        return None;
-    }
-    Some((remote.to_string(), branch.to_string()))
-}
-
-fn normalize_local_branch_ref(raw: &str) -> String {
-    raw.trim()
-        .trim_start_matches("refs/heads/")
-        .trim()
-        .to_string()
-}
-
-fn normalize_remote_target_branch(remote: &str, raw: &str) -> String {
-    let trimmed = raw.trim();
-    let without_refs = trimmed
-        .strip_prefix("refs/remotes/")
-        .or_else(|| trimmed.strip_prefix("remotes/"))
-        .unwrap_or(trimmed);
-    let remote_prefix = format!("{remote}/");
-    without_refs
-        .strip_prefix(&remote_prefix)
-        .unwrap_or(without_refs)
-        .trim()
-        .to_string()
-}
-
-fn upstream_remote_and_branch(repo_root: &Path) -> Result<Option<(String, String)>, String> {
-    let repo = open_repository_at_root(repo_root)?;
-    let head = match repo.head() {
-        Ok(head) => head,
-        Err(_) => return Ok(None),
-    };
-    if !head.is_branch() {
-        return Ok(None);
-    }
-    let branch_name = match head.shorthand() {
-        Some(name) => name,
-        None => return Ok(None),
-    };
-    let branch = repo
-        .find_branch(branch_name, BranchType::Local)
-        .map_err(|e| e.to_string())?;
-    let upstream_branch = match branch.upstream() {
-        Ok(upstream) => upstream,
-        Err(_) => return Ok(None),
-    };
-    let upstream_ref = upstream_branch.get();
-    let upstream_name = upstream_ref.name().or_else(|| upstream_ref.shorthand());
-    Ok(upstream_name.and_then(parse_upstream_ref))
-}
-
-fn current_local_branch(repo_root: &Path) -> Result<Option<String>, String> {
-    let repo = open_repository_at_root(repo_root)?;
-    let head = match repo.head() {
-        Ok(head) => head,
-        Err(_) => return Ok(None),
-    };
-    if !head.is_branch() {
-        return Ok(None);
-    }
-    Ok(head
-        .shorthand()
-        .map(normalize_local_branch_ref)
-        .filter(|name| !name.is_empty()))
-}
-
-fn split_csv_values(input: Option<String>) -> Vec<String> {
-    trim_optional(input)
-        .map(|value| {
-            value
-                .split(',')
-                .map(str::trim)
-                .filter(|entry| !entry.is_empty())
-                .map(ToOwned::to_owned)
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn build_gerrit_push_suffix(
-    topic: Option<String>,
-    reviewers: Option<String>,
-    cc: Option<String>,
-) -> String {
-    let mut params = Vec::new();
-    if let Some(topic_name) = trim_optional(topic) {
-        params.push(format!("topic={topic_name}"));
-    }
-    for reviewer in split_csv_values(reviewers) {
-        params.push(format!("r={reviewer}"));
-    }
-    for cc_member in split_csv_values(cc) {
-        params.push(format!("cc={cc_member}"));
-    }
-    params.join(",")
-}
-
-async fn push_with_upstream(repo_root: &Path) -> Result<(), String> {
-    let upstream = upstream_remote_and_branch(repo_root)?;
-    if let Some((remote, branch)) = upstream {
-        let refspec = format!("HEAD:{branch}");
-        return run_git_command(repo_root, &["push", remote.as_str(), refspec.as_str()]).await;
-    }
-    run_git_command(repo_root, &["push"]).await
-}
-
-async fn push_with_options(
-    repo_root: &Path,
-    remote: Option<String>,
-    branch: Option<String>,
-    force_with_lease: bool,
-    push_tags: bool,
-    run_hooks: bool,
-    push_to_gerrit: bool,
-    topic: Option<String>,
-    reviewers: Option<String>,
-    cc: Option<String>,
-) -> Result<(), String> {
-    let mut args = vec!["push".to_string()];
-    if !run_hooks {
-        args.push("--no-verify".to_string());
-    }
-    if force_with_lease {
-        args.push("--force-with-lease".to_string());
-    }
-    if push_tags {
-        args.push("--follow-tags".to_string());
-    }
-
-    let explicit_remote = trim_optional(remote);
-    let explicit_branch = trim_optional(branch)
-        .map(|value| normalize_local_branch_ref(&value))
-        .filter(|value| !value.is_empty());
-    let current_branch = current_local_branch(repo_root)?;
-    let target_branch = explicit_branch.or(current_branch);
-
-    if push_to_gerrit {
-        let target_remote = explicit_remote
-            .or_else(|| {
-                upstream_remote_and_branch(repo_root)
-                    .ok()
-                    .flatten()
-                    .map(|(name, _)| name)
-            })
-            .unwrap_or_else(|| "origin".to_string());
-        let target_branch =
-            target_branch.ok_or_else(|| "Branch is required for Gerrit push.".to_string())?;
-
-        let mut refspec = format!("HEAD:refs/for/{target_branch}");
-        let suffix = build_gerrit_push_suffix(topic, reviewers, cc);
-        if !suffix.is_empty() {
-            refspec.push('%');
-            refspec.push_str(&suffix);
-        }
-        args.push(target_remote);
-        args.push(refspec);
-        let command: Vec<&str> = args.iter().map(String::as_str).collect();
-        return run_git_command(repo_root, &command).await;
-    }
-
-    if explicit_remote.is_none() && target_branch.is_none() {
-        if !force_with_lease && !push_tags && run_hooks {
-            return push_with_upstream(repo_root).await;
-        }
-        let command: Vec<&str> = args.iter().map(String::as_str).collect();
-        return run_git_command(repo_root, &command).await;
-    }
-
-    let target_remote = explicit_remote
-        .or_else(|| {
-            upstream_remote_and_branch(repo_root)
-                .ok()
-                .flatten()
-                .map(|(name, _)| name)
-        })
-        .unwrap_or_else(|| "origin".to_string());
-    args.push(target_remote);
-    if let Some(target_branch) = target_branch {
-        args.push(format!("HEAD:{target_branch}"));
-    }
-
-    let command: Vec<&str> = args.iter().map(String::as_str).collect();
-    run_git_command(repo_root, &command).await
-}
-
-fn status_for_index(status: Status) -> Option<&'static str> {
-    if status.contains(Status::INDEX_NEW) {
-        Some("A")
-    } else if status.contains(Status::INDEX_MODIFIED) {
-        Some("M")
-    } else if status.contains(Status::INDEX_DELETED) {
-        Some("D")
-    } else if status.contains(Status::INDEX_RENAMED) {
-        Some("R")
-    } else if status.contains(Status::INDEX_TYPECHANGE) {
-        Some("T")
-    } else {
-        None
-    }
-}
-
-fn status_for_workdir(status: Status) -> Option<&'static str> {
-    if status.contains(Status::WT_NEW) {
-        Some("A")
-    } else if status.contains(Status::WT_MODIFIED) {
-        Some("M")
-    } else if status.contains(Status::WT_DELETED) {
-        Some("D")
-    } else if status.contains(Status::WT_RENAMED) {
-        Some("R")
-    } else if status.contains(Status::WT_TYPECHANGE) {
-        Some("T")
-    } else {
-        None
-    }
-}
-
-fn status_for_delta(status: git2::Delta) -> &'static str {
-    match status {
-        git2::Delta::Added => "A",
-        git2::Delta::Modified => "M",
-        git2::Delta::Deleted => "D",
-        git2::Delta::Renamed => "R",
-        git2::Delta::Typechange => "T",
-        _ => "M",
-    }
-}
-
-fn build_combined_diff(diff: &git2::Diff) -> String {
-    let mut combined_diff = String::new();
-    for (index, delta) in diff.deltas().enumerate() {
-        let path = delta.new_file().path().or_else(|| delta.old_file().path());
-        let Some(path) = path else {
-            continue;
-        };
-        let patch = match git2::Patch::from_diff(diff, index) {
-            Ok(patch) => patch,
-            Err(_) => continue,
-        };
-        let Some(mut patch) = patch else {
-            continue;
-        };
-        let content = match diff_patch_to_string(&mut patch) {
-            Ok(content) => content,
-            Err(_) => continue,
-        };
-        if content.trim().is_empty() {
-            continue;
-        }
-        if !combined_diff.is_empty() {
-            combined_diff.push_str("\n\n");
-        }
-        combined_diff.push_str(&format!("=== {} ===\n", path.display()));
-        combined_diff.push_str(&content);
-    }
-    combined_diff
-}
-
-fn collect_index_diff(
-    repo: &Repository,
-    head_tree: Option<&git2::Tree<'_>>,
-    pathspecs: Option<&[String]>,
-) -> Result<String, String> {
-    if matches!(pathspecs, Some(paths) if paths.is_empty()) {
-        return Ok(String::new());
-    }
-
-    let mut options = DiffOptions::new();
-    if let Some(paths) = pathspecs {
-        for path in paths {
-            options.pathspec(path);
-        }
-    }
-
-    let index = repo.index().map_err(|e| e.to_string())?;
-    let mut diff = match head_tree {
-        Some(tree) => repo
-            .diff_tree_to_index(Some(tree), Some(&index), Some(&mut options))
-            .map_err(|e| e.to_string())?,
-        None => repo
-            .diff_tree_to_index(None, Some(&index), Some(&mut options))
-            .map_err(|e| e.to_string())?,
-    };
-    find_git_diff_renames(&mut diff).map_err(|e| e.to_string())?;
-
-    Ok(build_combined_diff(&diff))
-}
-
-fn collect_worktree_diff(
-    repo: &Repository,
-    head_tree: Option<&git2::Tree<'_>>,
-    pathspecs: Option<&[String]>,
-) -> Result<String, String> {
-    if matches!(pathspecs, Some(paths) if paths.is_empty()) {
-        return Ok(String::new());
-    }
-
-    let mut options = DiffOptions::new();
-    options
-        .include_untracked(true)
-        .recurse_untracked_dirs(true)
-        .show_untracked_content(true);
-    if let Some(paths) = pathspecs {
-        for path in paths {
-            options.pathspec(path);
-        }
-    }
-
-    let mut diff = match head_tree {
-        Some(tree) => repo
-            .diff_tree_to_workdir_with_index(Some(tree), Some(&mut options))
-            .map_err(|e| e.to_string())?,
-        None => repo
-            .diff_tree_to_workdir_with_index(None, Some(&mut options))
-            .map_err(|e| e.to_string())?,
-    };
-    find_git_diff_renames(&mut diff).map_err(|e| e.to_string())?;
-
-    Ok(build_combined_diff(&diff))
-}
-
-fn collect_workspace_diff(repo_root: &Path) -> Result<String, String> {
-    let repo = open_repository_at_root(repo_root)?;
-    let head_tree = repo.head().ok().and_then(|head| head.peel_to_tree().ok());
-
-    let staged_diff = collect_index_diff(&repo, head_tree.as_ref(), None)?;
-    if !staged_diff.trim().is_empty() {
-        return Ok(staged_diff);
-    }
-
-    collect_worktree_diff(&repo, head_tree.as_ref(), None)
-}
-
-#[derive(Debug, Default, PartialEq, Eq)]
-struct CommitScopeDiffPlan {
-    index_paths: Vec<String>,
-    worktree_only_paths: Vec<String>,
-}
-
-fn normalize_commit_scope_path(path: &str) -> String {
-    normalize_git_path(path).trim_matches('/').to_string()
-}
-
-fn register_commit_scope_identity(
-    identities_by_path: &mut HashMap<String, Vec<String>>,
-    identity: &GitStatusPathIdentity,
-) {
-    let mut action_paths = Vec::new();
-    if let Some(old_path) = identity.old_path.as_ref() {
-        action_paths.push(old_path.clone());
-    }
-    if !action_paths.contains(&identity.path) {
-        action_paths.push(identity.path.clone());
-    }
-    for alias in &action_paths {
-        let normalized_alias = normalize_commit_scope_path(alias);
-        if !normalized_alias.is_empty() {
-            identities_by_path
-                .entry(normalized_alias)
-                .or_insert_with(|| action_paths.clone());
-        }
-    }
-}
-
-fn build_commit_scope_diff_plan(
-    repo: &Repository,
-    selected_paths: &[String],
-) -> Result<CommitScopeDiffPlan, String> {
-    let mut status_options = StatusOptions::new();
-    status_options
-        .include_untracked(true)
-        .recurse_untracked_dirs(true)
-        .renames_head_to_index(true)
-        .renames_index_to_workdir(true)
-        .include_ignored(false);
-
-    let statuses = repo
-        .statuses(Some(&mut status_options))
-        .map_err(|e| format!("failed to read git status for commit scope: {e}"))?;
-
-    let mut staged_by_normalized_path = HashMap::new();
-    let mut unstaged_by_normalized_path = HashMap::new();
-
-    for entry in statuses.iter() {
-        let status = entry.status();
-        if status.intersects(
-            Status::INDEX_NEW
-                | Status::INDEX_MODIFIED
-                | Status::INDEX_DELETED
-                | Status::INDEX_RENAMED
-                | Status::INDEX_TYPECHANGE,
-        ) {
-            if let Some(identity) = git_status_path_identity(&entry, GitStatusLayer::Index) {
-                register_commit_scope_identity(&mut staged_by_normalized_path, &identity);
-            }
-        }
-
-        if status.intersects(
-            Status::WT_NEW
-                | Status::WT_MODIFIED
-                | Status::WT_DELETED
-                | Status::WT_RENAMED
-                | Status::WT_TYPECHANGE,
-        ) {
-            if let Some(identity) = git_status_path_identity(&entry, GitStatusLayer::Workdir) {
-                register_commit_scope_identity(&mut unstaged_by_normalized_path, &identity);
-            }
-        }
-    }
-
-    let mut index_paths = Vec::new();
-    let mut worktree_only_paths = Vec::new();
-    let mut seen_index_paths = HashSet::new();
-    let mut seen_worktree_paths = HashSet::new();
-
-    for selected_path in selected_paths {
-        let normalized_path = normalize_commit_scope_path(selected_path);
-        if normalized_path.is_empty() {
-            continue;
-        }
-
-        if let Some(raw_paths) = staged_by_normalized_path.get(&normalized_path) {
-            for raw_path in raw_paths {
-                if seen_index_paths.insert(raw_path.clone()) {
-                    index_paths.push(raw_path.clone());
-                }
-            }
-            continue;
-        }
-
-        if let Some(raw_paths) = unstaged_by_normalized_path.get(&normalized_path) {
-            for raw_path in raw_paths {
-                if seen_worktree_paths.insert(raw_path.clone()) {
-                    worktree_only_paths.push(raw_path.clone());
-                }
-            }
-        }
-    }
-
-    Ok(CommitScopeDiffPlan {
-        index_paths,
-        worktree_only_paths,
-    })
-}
-
-fn collect_commit_scope_diff(
-    repo_root: &Path,
-    selected_paths: Option<&[String]>,
-) -> Result<String, String> {
-    let Some(explicit_selected_paths) = selected_paths else {
-        return collect_workspace_diff(repo_root);
-    };
-    if explicit_selected_paths.is_empty() {
-        return Ok(String::new());
-    }
-
-    let repo = open_repository_at_root(repo_root)?;
-    let head_tree = repo.head().ok().and_then(|head| head.peel_to_tree().ok());
-    let plan = build_commit_scope_diff_plan(&repo, explicit_selected_paths)?;
-
-    let staged_diff = collect_index_diff(&repo, head_tree.as_ref(), Some(&plan.index_paths))?;
-    let worktree_diff =
-        collect_worktree_diff(&repo, head_tree.as_ref(), Some(&plan.worktree_only_paths))?;
-
-    let mut segments = Vec::new();
-    if !staged_diff.trim().is_empty() {
-        segments.push(staged_diff);
-    }
-    if !worktree_diff.trim().is_empty() {
-        segments.push(worktree_diff);
-    }
-
-    Ok(segments.join("\n\n"))
-}
-
-fn github_repo_from_path(path: &Path) -> Result<String, String> {
-    let repo = open_repository_at_root(path)?;
-    let remotes = repo.remotes().map_err(|e| e.to_string())?;
-    let name = if remotes.iter().any(|remote| remote == Some("origin")) {
-        "origin".to_string()
-    } else {
-        remotes.iter().flatten().next().unwrap_or("").to_string()
-    };
-    if name.is_empty() {
-        return Err("No git remote configured.".to_string());
-    }
-    let remote = repo.find_remote(&name).map_err(|e| e.to_string())?;
-    let remote_url = remote.url().ok_or("Remote has no URL configured.")?;
-    parse_github_repo(remote_url).ok_or("Remote is not a GitHub repository.".to_string())
-}
-
-fn parse_patch_diff_entries(diff: &str) -> Vec<GitCommitDiff> {
-    let mut entries = Vec::new();
-    let mut current_lines: Vec<&str> = Vec::new();
-    let mut current_old_path: Option<String> = None;
-    let mut current_new_path: Option<String> = None;
-    let mut current_status: Option<String> = None;
-
-    let finalize = |lines: &Vec<&str>,
-                    old_path: &Option<String>,
-                    new_path: &Option<String>,
-                    status: &Option<String>,
-                    results: &mut Vec<GitCommitDiff>| {
-        if lines.is_empty() {
-            return;
-        }
-        let diff_text = lines.join("\n");
-        if diff_text.trim().is_empty() {
-            return;
-        }
-        let status_value = status.clone().unwrap_or_else(|| "M".to_string());
-        let path = if status_value == "D" {
-            old_path.clone().unwrap_or_default()
-        } else {
-            new_path
-                .clone()
-                .or_else(|| old_path.clone())
-                .unwrap_or_default()
-        };
-        if path.is_empty() {
-            return;
-        }
-        results.push(GitCommitDiff {
-            path: normalize_git_path(&path),
-            status: status_value,
-            diff: diff_text,
-            is_binary: false,
-            is_image: false,
-            old_image_data: None,
-            new_image_data: None,
-            old_image_mime: None,
-            new_image_mime: None,
-        });
-    };
-
-    for line in diff.lines() {
-        if line.starts_with("diff --git ") {
-            finalize(
-                &current_lines,
-                &current_old_path,
-                &current_new_path,
-                &current_status,
-                &mut entries,
-            );
-            current_lines = vec![line];
-            current_old_path = None;
-            current_new_path = None;
-            current_status = None;
-
-            let rest = line.trim_start_matches("diff --git ").trim();
-            let mut parts = rest.split_whitespace();
-            let old_part = parts.next().unwrap_or("").trim_start_matches("a/");
-            let new_part = parts.next().unwrap_or("").trim_start_matches("b/");
-            if !old_part.is_empty() {
-                current_old_path = Some(old_part.to_string());
-            }
-            if !new_part.is_empty() {
-                current_new_path = Some(new_part.to_string());
-            }
-            continue;
-        }
-        if line.starts_with("new file mode ") {
-            current_status = Some("A".to_string());
-        } else if line.starts_with("deleted file mode ") {
-            current_status = Some("D".to_string());
-        } else if line.starts_with("rename from ") {
-            current_status = Some("R".to_string());
-            let path = line.trim_start_matches("rename from ").trim();
-            if !path.is_empty() {
-                current_old_path = Some(path.to_string());
-            }
-        } else if line.starts_with("rename to ") {
-            current_status = Some("R".to_string());
-            let path = line.trim_start_matches("rename to ").trim();
-            if !path.is_empty() {
-                current_new_path = Some(path.to_string());
-            }
-        }
-        current_lines.push(line);
-    }
-
-    finalize(
-        &current_lines,
-        &current_old_path,
-        &current_new_path,
-        &current_status,
-        &mut entries,
-    );
-
-    entries
-}
-
-fn parse_pr_diff(diff: &str) -> Vec<GitHubPullRequestDiff> {
-    parse_patch_diff_entries(diff)
-        .into_iter()
-        .map(|entry| GitHubPullRequestDiff {
-            path: entry.path,
-            status: entry.status,
-            diff: entry.diff,
-        })
-        .collect()
-}
-
-#[derive(Debug)]
-struct TokenIsolatedCommandOutput {
-    success: bool,
-    command: String,
-    stdout: String,
-    stderr: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct GhExistingPrEntry {
-    number: u64,
-    title: String,
-    url: String,
-    state: String,
-    #[serde(rename = "headRefName")]
-    head_ref_name: String,
-    #[serde(rename = "baseRefName")]
-    base_ref_name: String,
-}
-
-fn shell_escape_for_display(value: &str) -> String {
-    if value.is_empty() {
-        return "''".to_string();
-    }
-    if value.chars().all(|ch| {
-        ch.is_ascii_alphanumeric() || matches!(ch, '/' | '_' | '-' | '.' | ':' | '@' | '=')
-    }) {
-        return value.to_string();
-    }
-    format!("'{}'", value.replace('\'', "'\\''"))
-}
-
-fn build_token_isolated_command_display(program: &str, args: &[String]) -> String {
-    let mut rendered = vec!["env -u GH_TOKEN -u GITHUB_TOKEN".to_string()];
-    rendered.push(shell_escape_for_display(program));
-    rendered.extend(args.iter().map(|value| shell_escape_for_display(value)));
-    rendered.join(" ")
-}
-
-fn summarize_command_failure(output: &TokenIsolatedCommandOutput) -> String {
-    let stderr = output.stderr.trim();
-    if !stderr.is_empty() {
-        return stderr.to_string();
-    }
-    let stdout = output.stdout.trim();
-    if !stdout.is_empty() {
-        return stdout.to_string();
-    }
-    "Command failed without stderr/stdout output.".to_string()
-}
-
-fn truncate_debug_text(raw: &str, max_len: usize) -> String {
-    if raw.chars().count() <= max_len {
-        return raw.to_string();
-    }
-    raw.chars().take(max_len).collect::<String>() + " ...[truncated]"
-}
-
-async fn run_token_isolated_command(
-    repo_root: &Path,
-    program: &str,
-    args: &[String],
-    extra_env: &[(&str, &str)],
-) -> Result<TokenIsolatedCommandOutput, String> {
-    let mut command = if program == "git" {
-        crate::utils::async_command(
-            resolve_git_binary().map_err(|e| format!("Failed to run git: {e}"))?,
-        )
-    } else {
-        crate::utils::async_command(program)
-    };
-    command
-        .args(args)
-        .current_dir(repo_root)
-        .env("PATH", git_env_path())
-        .env_remove("GH_TOKEN")
-        .env_remove("GITHUB_TOKEN")
-        .env("GIT_TERMINAL_PROMPT", "0")
-        .env("GCM_INTERACTIVE", "never")
-        .stdin(Stdio::null())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-    for (key, value) in extra_env {
-        command.env(key, value);
-    }
-
-    let output = match timeout(
-        Duration::from_secs(GIT_COMMAND_TIMEOUT_SECS),
-        command.output(),
-    )
-    .await
-    {
-        Ok(result) => result.map_err(|error| {
-            if program == "gh" {
-                format!("Failed to run gh command: {error}. Ensure GitHub CLI (gh) is installed.")
-            } else {
-                format!("Failed to run {program} command: {error}")
-            }
-        })?,
-        Err(_) => {
-            return Err(format!(
-                "Command timed out after {GIT_COMMAND_TIMEOUT_SECS}s: {}",
-                build_token_isolated_command_display(program, args)
-            ));
-        }
-    };
-
-    Ok(TokenIsolatedCommandOutput {
-        success: output.status.success(),
-        command: build_token_isolated_command_display(program, args),
-        stdout: String::from_utf8_lossy(&output.stdout).trim().to_string(),
-        stderr: String::from_utf8_lossy(&output.stderr).trim().to_string(),
-    })
-}
-
-fn is_http2_transport_error(raw: &str) -> bool {
-    let normalized = raw.to_lowercase();
-    normalized.contains("http2 framing layer")
-        || normalized.contains("http/2 stream")
-        || normalized.contains("stream 0 was not closed cleanly")
-}
-
-fn is_auth_related_error(raw: &str) -> bool {
-    let normalized = raw.to_lowercase();
-    normalized.contains("403")
-        || normalized.contains("authentication failed")
-        || normalized.contains("permission denied")
-        || normalized.contains("resource not accessible by personal access token")
-        || normalized.contains("requires authentication")
-        || normalized.contains("not logged into any github hosts")
-}
-
-fn is_network_related_error(raw: &str) -> bool {
-    let normalized = raw.to_lowercase();
-    normalized.contains("failed to connect")
-        || normalized.contains("could not resolve host")
-        || normalized.contains("timed out")
-        || normalized.contains("connection reset")
-        || normalized.contains("network is unreachable")
-}
-
-fn parse_repo_owner(repo: &str) -> Option<String> {
-    repo.split('/')
-        .next()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn resolve_remote_repo(repo: &Repository, remote_name: &str) -> Option<String> {
-    let remote = repo.find_remote(remote_name).ok()?;
-    let remote_url = remote.url()?;
-    parse_github_repo(remote_url)
-}
-
-fn infer_remote_head_branch(repo: &Repository, remote_name: &str) -> Option<String> {
-    let head_ref = format!("refs/remotes/{remote_name}/HEAD");
-    let reference = repo.find_reference(&head_ref).ok()?;
-    let symbolic_target = reference.symbolic_target()?;
-    let prefix = format!("refs/remotes/{remote_name}/");
-    symbolic_target
-        .strip_prefix(prefix.as_str())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn default_pr_description(base_branch: &str, head_branch: &str) -> String {
-    format!(
-        "## 背景\n- 从 `{base_branch}` 合并到 `{head_branch}`。\n\n## 改动点\n- \n\n## 验证\n- [ ] npm run typecheck\n- [ ] npm run lint"
-    )
-}
-
-fn build_workflow_stages() -> Vec<GitPrWorkflowStage> {
-    vec![
-        GitPrWorkflowStage {
-            key: "precheck".to_string(),
-            status: "pending".to_string(),
-            detail: "Waiting for precheck.".to_string(),
-            command: None,
-            stdout: None,
-            stderr: None,
-        },
-        GitPrWorkflowStage {
-            key: "push".to_string(),
-            status: "pending".to_string(),
-            detail: "Waiting for push.".to_string(),
-            command: None,
-            stdout: None,
-            stderr: None,
-        },
-        GitPrWorkflowStage {
-            key: "create".to_string(),
-            status: "pending".to_string(),
-            detail: "Waiting for PR creation.".to_string(),
-            command: None,
-            stdout: None,
-            stderr: None,
-        },
-        GitPrWorkflowStage {
-            key: "comment".to_string(),
-            status: "pending".to_string(),
-            detail: "Waiting for optional comment.".to_string(),
-            command: None,
-            stdout: None,
-            stderr: None,
-        },
-    ]
-}
-
-fn update_workflow_stage(
-    stages: &mut [GitPrWorkflowStage],
-    key: &str,
-    status: &str,
-    detail: String,
-    command: Option<String>,
-    stdout: Option<String>,
-    stderr: Option<String>,
-) {
-    if let Some(stage) = stages.iter_mut().find(|entry| entry.key == key) {
-        stage.status = status.to_string();
-        stage.detail = detail;
-        stage.command = command;
-        stage.stdout = stdout;
-        stage.stderr = stderr;
-    }
-}
-
-fn stage_error_category_and_hint(stage_key: &str, raw: &str) -> (String, String) {
-    if stage_key == "precheck" {
-        if raw.to_lowercase().contains("gh") && raw.to_lowercase().contains("not found") {
-            return (
-                "gh-not-installed".to_string(),
-                "Install GitHub CLI and run `gh auth login` first.".to_string(),
-            );
-        }
-        if is_auth_related_error(raw) {
-            return (
-                "gh-auth-missing".to_string(),
-                "Run `env -u GH_TOKEN -u GITHUB_TOKEN gh auth status -h github.com` and finish login.".to_string(),
-            );
-        }
-        if raw.to_lowercase().contains("range gate") {
-            return (
-                "range-abnormal".to_string(),
-                "Review changed files against upstream base, then re-run after rebasing/fixing scope.".to_string(),
-            );
-        }
-    }
-    if stage_key == "push" {
-        if is_http2_transport_error(raw) {
-            return (
-                "push-http2".to_string(),
-                "Retry with HTTP/1.1 fallback: `git -c http.version=HTTP/1.1 push ...`."
-                    .to_string(),
-            );
-        }
-        if is_auth_related_error(raw) {
-            return (
-                "push-auth".to_string(),
-                "Verify fork push permission and run with token-isolated env (`env -u GH_TOKEN -u GITHUB_TOKEN`).".to_string(),
-            );
-        }
-        if is_network_related_error(raw) {
-            return (
-                "push-network".to_string(),
-                "Check network/proxy connectivity to github.com:443, then retry.".to_string(),
-            );
-        }
-    }
-    if stage_key == "create" {
-        if is_auth_related_error(raw) {
-            return (
-                "create-pr-auth".to_string(),
-                "Use `env -u GH_TOKEN -u GITHUB_TOKEN` and ensure `gh auth status` is healthy."
-                    .to_string(),
-            );
-        }
-        if is_network_related_error(raw) {
-            return (
-                "create-pr-network".to_string(),
-                "Network seems unstable. Retry once after validating GitHub connectivity."
-                    .to_string(),
-            );
-        }
-    }
-    (
-        "unknown".to_string(),
-        "Check stage stderr and retry.".to_string(),
-    )
-}
-
-fn extract_pr_url(raw: &str) -> Option<String> {
-    raw.split_whitespace()
-        .find(|token| token.starts_with("https://") && token.contains("/pull/"))
-        .map(|token| {
-            token
-                .trim()
-                .trim_matches('\'')
-                .trim_matches('"')
-                .trim_end_matches('.')
-                .to_string()
-        })
-}
-
-fn extract_pr_number_from_url(url: &str) -> Option<u64> {
-    let pull_segment = url.split("/pull/").nth(1)?;
-    let number_text = pull_segment
-        .split(['/', '?', '#'])
-        .next()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())?;
-    number_text.parse::<u64>().ok()
-}
-
-fn build_failed_pr_workflow_result(
-    stages: Vec<GitPrWorkflowStage>,
-    stage_key: &str,
-    raw_error: String,
-    retry_command: Option<String>,
-) -> GitPrWorkflowResult {
-    let (category, hint) = stage_error_category_and_hint(stage_key, &raw_error);
-    GitPrWorkflowResult {
-        ok: false,
-        status: "failed".to_string(),
-        message: raw_error,
-        error_category: Some(category),
-        next_action_hint: Some(hint),
-        pr_url: None,
-        pr_number: None,
-        existing_pr: None,
-        retry_command,
-        range_gate: None,
-        stages,
-    }
-}
-
-fn build_existing_pr_workflow_result(
-    stages: Vec<GitPrWorkflowStage>,
-    existing_pr: GitPrExistingPullRequest,
-) -> GitPrWorkflowResult {
-    GitPrWorkflowResult {
-        ok: true,
-        status: "existing".to_string(),
-        message: format!(
-            "Existing PR detected: #{} {}",
-            existing_pr.number, existing_pr.title
-        ),
-        error_category: None,
-        next_action_hint: Some(
-            "Open the existing PR and continue updates on the same branch.".to_string(),
-        ),
-        pr_url: Some(existing_pr.url.clone()),
-        pr_number: Some(existing_pr.number),
-        existing_pr: Some(existing_pr),
-        retry_command: None,
-        range_gate: None,
-        stages,
-    }
-}
-
-fn build_success_pr_workflow_result(
-    stages: Vec<GitPrWorkflowStage>,
-    pr_url: String,
-    pr_number: Option<u64>,
-    message: String,
-) -> GitPrWorkflowResult {
-    GitPrWorkflowResult {
-        ok: true,
-        status: "success".to_string(),
-        message,
-        error_category: None,
-        next_action_hint: None,
-        pr_url: Some(pr_url),
-        pr_number,
-        existing_pr: None,
-        retry_command: None,
-        range_gate: None,
-        stages,
-    }
-}
-
 mod commands;
 pub(crate) use commands::*;
 
+mod diff_collect;
+mod pr_workflow_helpers;
+mod push;
+mod remote_forward;
+
+pub(crate) use diff_collect::*;
+pub(crate) use pr_workflow_helpers::*;
+pub(crate) use push::*;
+pub(crate) use remote_forward::*;
+
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::types::{GitBlameHunk, GitLogEntry};
-    use serde_json::Value;
-    use std::fs;
-    use validation::validate_local_branch_name;
-
-    #[test]
-    fn git_remote_forwarding_matrix_has_unique_daemon_methods() {
-        let mut methods = GIT_REMOTE_FORWARDING_MATRIX
-            .iter()
-            .map(|entry| entry.method)
-            .collect::<Vec<_>>();
-        methods.sort_unstable();
-        methods.dedup();
-        assert_eq!(
-            methods.len(),
-            GIT_REMOTE_FORWARDING_MATRIX.len(),
-            "Git remote forwarding matrix contains duplicate methods"
-        );
-
-        let daemon_dispatch = include_str!("../bin/cc_gui_daemon.rs");
-        let mut categories = HashSet::new();
-        for entry in GIT_REMOTE_FORWARDING_MATRIX {
-            categories.insert(entry.category);
-            assert_eq!(entry.forwarding, "implemented");
-            assert!(!entry.desktop_module.trim().is_empty());
-            assert!(!entry.coverage.trim().is_empty());
-            assert!(
-                daemon_dispatch.contains(&format!("\"{}\" =>", entry.daemon_dispatch)),
-                "daemon dispatch is missing Git remote method {}",
-                entry.daemon_dispatch
-            );
-        }
-        for category in ["read", "history", "write", "branch", "worktree", "github"] {
-            assert!(
-                categories.contains(category),
-                "Git remote forwarding matrix is missing category {category}"
-            );
-        }
-    }
-
-    #[test]
-    fn pull_request_content_generation_remote_mode_fails_closed() {
-        let command_source = include_str!("commands.rs");
-        let command = command_source
-            .split("pub(crate) async fn generate_pull_request_content(")
-            .nth(1)
-            .expect("PR content generation command");
-        assert!(command.contains("PR content generation is unavailable in remote mode"));
-        assert!(!command.contains("\"generate_pull_request_content\","));
-    }
-
-    #[test]
-    fn git_file_blame_daemon_contract_preserves_payload_and_response_shape() {
-        let daemon_dispatch = include_str!("../bin/cc_gui_daemon.rs");
-        let blame_arm = daemon_dispatch
-            .split("\"get_git_file_blame\" => {")
-            .nth(1)
-            .and_then(|tail| tail.split("\"get_git_log\" => {").next())
-            .expect("daemon blame dispatch arm");
-        for payload_field in ["workspaceId", "path", "repositoryRoot"] {
-            assert!(
-                blame_arm.contains(payload_field),
-                "daemon blame dispatch is missing payload field {payload_field}"
-            );
-        }
-        assert!(blame_arm.contains("serde_json::to_value(response)"));
-
-        let response = GitFileBlameResponse {
-            path: "src/main.rs".to_string(),
-            head_sha: "abc123".to_string(),
-            line_count: 2,
-            hunks: vec![GitBlameHunk {
-                start_line: 1,
-                line_count: 2,
-                commit_sha: "abc123".to_string(),
-                author: "Ada".to_string(),
-                authored_at: 1_700_000_000,
-                summary: "Initial commit".to_string(),
-                original_path: None,
-            }],
-        };
-        let value = serde_json::to_value(&response).expect("serialize blame response");
-        assert_eq!(value.get("headSha").and_then(Value::as_str), Some("abc123"));
-        assert_eq!(value.get("lineCount").and_then(Value::as_u64), Some(2));
-        let hunk = value
-            .get("hunks")
-            .and_then(Value::as_array)
-            .and_then(|hunks| hunks.first())
-            .expect("serialized blame hunk");
-        assert_eq!(hunk.get("startLine").and_then(Value::as_u64), Some(1));
-        assert_eq!(
-            hunk.get("authoredAt").and_then(Value::as_i64),
-            Some(1_700_000_000)
-        );
-        assert_eq!(
-            serde_json::from_value::<GitFileBlameResponse>(value)
-                .expect("deserialize blame response"),
-            response
-        );
-    }
-
-    #[test]
-    fn git_log_payload_budget_serializes_content_safe_metadata() {
-        let response = GitLogResponse {
-            total: 10,
-            entries: vec![GitLogEntry {
-                sha: "abc123".to_string(),
-                summary: "summary".to_string(),
-                author: "author".to_string(),
-                timestamp: 1,
-            }],
-            ahead: 0,
-            behind: 0,
-            ahead_entries: Vec::new(),
-            behind_entries: Vec::new(),
-            upstream: None,
-            payload_budget: Some(PayloadBudgetMetadata {
-                command: "get_git_log".to_string(),
-                surface_id: "git-history-log".to_string(),
-                item_count: 1,
-                estimated_bytes: 64,
-                partial: true,
-                truncated: true,
-                cache_state: ScanCacheState::Unsupported,
-                evidence_class: "proxy".to_string(),
-            }),
-        };
-
-        let value = serde_json::to_value(response).expect("serialize response");
-        let budget = value
-            .get("payloadBudget")
-            .and_then(|entry| entry.as_object())
-            .expect("payload budget object");
-        assert_eq!(
-            budget.get("command").and_then(|entry| entry.as_str()),
-            Some("get_git_log")
-        );
-        assert_eq!(
-            budget.get("itemCount").and_then(|entry| entry.as_u64()),
-            Some(1)
-        );
-        assert!(budget.get("absolutePath").is_none());
-        assert!(budget.get("prompt").is_none());
-        assert!(budget.get("toolOutput").is_none());
-    }
-
-    fn create_temp_repo() -> (PathBuf, Repository) {
-        let root = std::env::temp_dir().join(format!("moss-x-test-{}", uuid::Uuid::new_v4()));
-        fs::create_dir_all(&root).expect("create temp repo root");
-        let repo = Repository::init(&root).expect("init repo");
-        (root, repo)
-    }
-
-    async fn commit_all_with_message(repo_root: &Path, message: &str) {
-        run_git_command(repo_root, &["add", "-A"])
-            .await
-            .expect("stage files");
-        run_git_command(
-            repo_root,
-            &[
-                "-c",
-                "user.name=TestUser",
-                "-c",
-                "user.email=test@example.com",
-                "commit",
-                "-m",
-                message,
-            ],
-        )
-        .await
-        .expect("commit staged files");
-    }
-
-    fn assert_worktree_clean(repo_root: &Path) {
-        let repo = open_repository_at_root(repo_root).expect("open repo");
-        let mut status_options = StatusOptions::new();
-        status_options
-            .include_untracked(true)
-            .recurse_untracked_dirs(true)
-            .include_ignored(false);
-        let statuses = repo
-            .statuses(Some(&mut status_options))
-            .expect("collect statuses");
-        assert!(
-            statuses.is_empty(),
-            "expected clean worktree, found {} entries",
-            statuses.len()
-        );
-    }
-
-    #[tokio::test]
-    async fn checkout_roundtrip_between_divergent_branches_stays_clean() {
-        let (root, _repo) = create_temp_repo();
-
-        fs::write(root.join("shared.txt"), "main base\n").expect("write initial main file");
-        commit_all_with_message(&root, "init main").await;
-        run_git_command(&root, &["branch", "-M", "main"])
-            .await
-            .expect("rename default branch to main");
-
-        run_git_command(&root, &["checkout", "-b", "feature/divergent"])
-            .await
-            .expect("create feature branch");
-        fs::remove_file(root.join("shared.txt")).expect("remove shared file on feature branch");
-        fs::write(root.join("feature-only.txt"), "feature branch content\n")
-            .expect("write feature-only file");
-        commit_all_with_message(&root, "feature commit").await;
-
-        run_git_command(&root, &["checkout", "main"])
-            .await
-            .expect("switch back to main");
-        fs::write(root.join("shared.txt"), "main updated\n").expect("rewrite shared file on main");
-        fs::write(root.join("main-only.txt"), "main branch content\n")
-            .expect("write main-only file");
-        commit_all_with_message(&root, "main commit").await;
-
-        for target in ["feature/divergent", "main", "feature/divergent", "main"] {
-            run_git_command(&root, &["checkout", target])
-                .await
-                .unwrap_or_else(|error| panic!("checkout {target} failed: {error}"));
-            assert_worktree_clean(&root);
-        }
-    }
-
-    #[tokio::test]
-    async fn file_history_follows_root_file_rename() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("before.txt"), "first\n").expect("write original file");
-        commit_all_with_message(&root, "create original").await;
-        fs::rename(root.join("before.txt"), root.join("after.txt")).expect("rename file");
-        commit_all_with_message(&root, "rename file").await;
-        fs::write(root.join("after.txt"), "first\nsecond\n").expect("edit renamed file");
-        commit_all_with_message(&root, "edit renamed file").await;
-
-        let entries = crate::shared::git_core::list_file_history_entries(&root, None, "after.txt")
-            .await
-            .expect("list file history");
-
-        assert_eq!(entries.len(), 3);
-        assert!(entries.iter().all(|entry| entry.oid.len() == 40));
-        assert_eq!(entries[0].path, "after.txt");
-        assert_eq!(entries[1].path, "after.txt");
-        assert_eq!(entries[2].path, "before.txt");
-
-        let option_like_ref =
-            crate::shared::git_core::list_file_history_oids(&root, Some("--no-walk"), "after.txt")
-                .await;
-        assert!(option_like_ref.is_err());
-    }
-
-    #[test]
-    fn file_history_rejects_invalid_paths() {
-        for path in [
-            "",
-            "/absolute.txt",
-            "../outside.txt",
-            "src/../../outside.txt",
-        ] {
-            assert!(crate::shared::git_core::normalize_file_history_path(path).is_err());
-        }
-        assert_eq!(
-            crate::shared::git_core::normalize_file_history_path("src\\main.rs")
-                .expect("normalize Windows path"),
-            "src/main.rs"
-        );
-    }
-
-    #[test]
-    fn file_history_path_participates_in_snapshot_identity() {
-        let first = crate::shared::git_core::build_git_history_snapshot_id(
-            "head",
-            None,
-            None,
-            None,
-            None,
-            None,
-            Some("packages/app"),
-            Some("src/first.ts"),
-        );
-        let second = crate::shared::git_core::build_git_history_snapshot_id(
-            "head",
-            None,
-            None,
-            None,
-            None,
-            None,
-            Some("packages/app"),
-            Some("src/second.ts"),
-        );
-        assert_ne!(first, second);
-    }
-
-    #[test]
-    fn collect_workspace_diff_prefers_staged_changes() {
-        let (root, repo) = create_temp_repo();
-        let file_path = root.join("staged.txt");
-        fs::write(&file_path, "staged\n").expect("write staged file");
-        let mut index = repo.index().expect("index");
-        index.add_path(Path::new("staged.txt")).expect("add path");
-        index.write().expect("write index");
-
-        let diff = collect_workspace_diff(&root).expect("collect diff");
-        assert!(diff.contains("staged.txt"));
-        assert!(diff.contains("staged"));
-    }
-
-    #[test]
-    fn collect_workspace_diff_falls_back_to_workdir() {
-        let (root, _repo) = create_temp_repo();
-        let file_path = root.join("unstaged.txt");
-        fs::write(&file_path, "unstaged\n").expect("write unstaged file");
-
-        let diff = collect_workspace_diff(&root).expect("collect diff");
-        assert!(diff.contains("unstaged.txt"));
-        assert!(diff.contains("unstaged"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_limits_selected_staged_files() {
-        let (root, repo) = create_temp_repo();
-        fs::write(root.join("selected.txt"), "selected\n").expect("write selected file");
-        fs::write(root.join("ignored.txt"), "ignored\n").expect("write ignored file");
-
-        let mut index = repo.index().expect("repo index");
-        index
-            .add_path(Path::new("selected.txt"))
-            .expect("stage selected file");
-        index
-            .add_path(Path::new("ignored.txt"))
-            .expect("stage ignored file");
-        index.write().expect("write index");
-
-        let selected_paths = vec!["selected.txt".to_string()];
-        let diff =
-            collect_commit_scope_diff(&root, Some(&selected_paths)).expect("collect scoped diff");
-        assert!(diff.contains("selected.txt"));
-        assert!(!diff.contains("ignored.txt"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_includes_selected_unstaged_only_file() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("selected.txt"), "selected\n").expect("write selected file");
-        fs::write(root.join("ignored.txt"), "ignored\n").expect("write ignored file");
-
-        let selected_paths = vec!["selected.txt".to_string()];
-        let diff =
-            collect_commit_scope_diff(&root, Some(&selected_paths)).expect("collect scoped diff");
-        assert!(diff.contains("selected.txt"));
-        assert!(!diff.contains("ignored.txt"));
-    }
-
-    #[tokio::test]
-    async fn collect_commit_scope_diff_uses_only_staged_portion_for_hybrid_path() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("hybrid.txt"), "before\n").expect("write initial file");
-        commit_all_with_message(&root, "init hybrid").await;
-
-        fs::write(root.join("hybrid.txt"), "staged only\n").expect("write staged content");
-        run_git_command(&root, &["add", "--", "hybrid.txt"])
-            .await
-            .expect("stage hybrid file");
-        fs::write(root.join("hybrid.txt"), "staged only\nunstaged extra\n")
-            .expect("write unstaged tail");
-
-        let selected_paths = vec!["hybrid.txt".to_string()];
-        let diff =
-            collect_commit_scope_diff(&root, Some(&selected_paths)).expect("collect scoped diff");
-        assert!(diff.contains("hybrid.txt"));
-        assert!(diff.contains("staged only"));
-        assert!(!diff.contains("unstaged extra"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_normalizes_windows_style_selected_paths() {
-        let (root, _repo) = create_temp_repo();
-        let nested_dir = root.join("src").join("feature");
-        fs::create_dir_all(&nested_dir).expect("create nested dir");
-        fs::write(nested_dir.join("file.ts"), "console.log('hi');\n").expect("write nested file");
-        fs::write(root.join("ignored.ts"), "console.log('ignored');\n")
-            .expect("write sibling file");
-
-        let selected_paths = vec!["src\\feature\\file.ts".to_string()];
-        let diff =
-            collect_commit_scope_diff(&root, Some(&selected_paths)).expect("collect scoped diff");
-        assert!(diff.contains("src/feature/file.ts"));
-        assert!(!diff.contains("ignored.ts"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_resolves_staged_rename_destination() {
-        let (root, repo) = create_temp_repo();
-        fs::write(root.join("before.txt"), "rename content\n").expect("write source file");
-        let mut index = repo.index().expect("repo index");
-        index
-            .add_path(Path::new("before.txt"))
-            .expect("stage source file");
-        let tree_id = index.write_tree().expect("write source tree");
-        let tree = repo.find_tree(tree_id).expect("find source tree");
-        let signature = git2::Signature::now("Test", "test@example.com").expect("signature");
-        repo.commit(Some("HEAD"), &signature, &signature, "init", &tree, &[])
-            .expect("commit source file");
-        drop(tree);
-
-        fs::rename(root.join("before.txt"), root.join("after.txt")).expect("rename source file");
-        let mut index = repo.index().expect("repo index");
-        index
-            .remove_path(Path::new("before.txt"))
-            .expect("remove source path");
-        index
-            .add_path(Path::new("after.txt"))
-            .expect("stage destination path");
-        index.write().expect("write renamed index");
-
-        let selected_paths = vec!["after.txt".to_string()];
-        let diff =
-            collect_commit_scope_diff(&root, Some(&selected_paths)).expect("collect rename diff");
-
-        assert!(diff.contains("before.txt"));
-        assert!(diff.contains("after.txt"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_keeps_staged_first_fallback_without_explicit_scope() {
-        let (root, repo) = create_temp_repo();
-        fs::write(root.join("staged.txt"), "staged\n").expect("write staged file");
-        fs::write(root.join("unstaged.txt"), "unstaged\n").expect("write unstaged file");
-
-        let mut index = repo.index().expect("repo index");
-        index
-            .add_path(Path::new("staged.txt"))
-            .expect("stage staged file");
-        index.write().expect("write index");
-
-        let diff = collect_commit_scope_diff(&root, None).expect("collect scoped diff");
-        assert!(diff.contains("staged.txt"));
-        assert!(!diff.contains("unstaged.txt"));
-    }
-
-    #[test]
-    fn collect_commit_scope_diff_returns_empty_for_explicit_empty_scope() {
-        let (root, repo) = create_temp_repo();
-        fs::write(root.join("staged.txt"), "staged\n").expect("write staged file");
-
-        let mut index = repo.index().expect("repo index");
-        index
-            .add_path(Path::new("staged.txt"))
-            .expect("stage staged file");
-        index.write().expect("write index");
-
-        let explicit_empty: Vec<String> = Vec::new();
-        let diff =
-            collect_commit_scope_diff(&root, Some(&explicit_empty)).expect("collect scoped diff");
-        assert!(diff.trim().is_empty());
-    }
-
-    #[test]
-    fn heavy_diff_path_guard_matches_lockfiles_and_generated_dirs() {
-        assert!(is_heavy_diff_path("pnpm-lock.yaml"));
-        assert!(is_heavy_diff_path(
-            "packages/web/node_modules/lodash/index.js"
-        ));
-        assert!(is_heavy_diff_path("dist/main.bundle.js"));
-        assert!(!is_heavy_diff_path("src/features/git/mod.rs"));
-    }
-
-    #[test]
-    fn truncate_diff_preview_respects_line_and_byte_budgets() {
-        let mut content = String::new();
-        for _ in 0..20 {
-            content.push_str("0123456789abcdef\n");
-        }
-        let trimmed = truncate_diff_preview(content, 4, 40);
-        assert!(trimmed.contains("[diff truncated for performance]"));
-        assert!(trimmed.lines().count() <= 6);
-        assert!(trimmed.len() <= 80);
-    }
-
-    #[test]
-    fn action_paths_for_file_expands_renames() {
-        let (root, repo) = create_temp_repo();
-        fs::write(root.join("a.txt"), "hello\n").expect("write file");
-
-        let mut index = repo.index().expect("repo index");
-        index.add_path(Path::new("a.txt")).expect("add path");
-        let tree_id = index.write_tree().expect("write tree");
-        let tree = repo.find_tree(tree_id).expect("find tree");
-        let sig = git2::Signature::now("Test", "test@example.com").expect("signature");
-        repo.commit(Some("HEAD"), &sig, &sig, "init", &tree, &[])
-            .expect("commit");
-
-        fs::rename(root.join("a.txt"), root.join("b.txt")).expect("rename file");
-
-        // Stage the rename so libgit2 reports it as an INDEX_RENAMED entry.
-        let mut index = repo.index().expect("repo index");
-        index
-            .remove_path(Path::new("a.txt"))
-            .expect("remove old path");
-        index.add_path(Path::new("b.txt")).expect("add new path");
-        index.write().expect("write index");
-
-        let paths = action_paths_for_file(&root, "b.txt", GitStatusLayer::Index);
-        assert_eq!(paths, vec!["a.txt".to_string(), "b.txt".to_string()]);
-    }
-
-    #[tokio::test]
-    async fn chained_rename_stage_and_unstage_mutate_the_intended_layer() {
-        let (stage_root, _stage_repo) = create_temp_repo();
-        fs::write(stage_root.join("a.txt"), "hello\n").expect("write staged fixture");
-        commit_all_with_message(&stage_root, "init staged fixture").await;
-        run_git_command(&stage_root, &["mv", "a.txt", "b.txt"])
-            .await
-            .expect("stage a to b rename");
-        fs::rename(stage_root.join("b.txt"), stage_root.join("c.txt"))
-            .expect("rename b to c in workdir");
-
-        for path in action_paths_for_file(&stage_root, "c.txt", GitStatusLayer::Workdir) {
-            run_git_command(&stage_root, &["add", "-A", "--", &path])
-                .await
-                .expect("stage workdir rename path");
-        }
-        let stage_index = open_repository_at_root(&stage_root)
-            .expect("open staged repo")
-            .index()
-            .expect("open staged index");
-        assert!(stage_index.get_path(Path::new("a.txt"), 0).is_none());
-        assert!(stage_index.get_path(Path::new("b.txt"), 0).is_none());
-        assert!(stage_index.get_path(Path::new("c.txt"), 0).is_some());
-
-        let (unstage_root, _unstage_repo) = create_temp_repo();
-        fs::write(unstage_root.join("a.txt"), "hello\n").expect("write unstaged fixture");
-        commit_all_with_message(&unstage_root, "init unstaged fixture").await;
-        run_git_command(&unstage_root, &["mv", "a.txt", "b.txt"])
-            .await
-            .expect("stage a to b rename");
-        fs::rename(unstage_root.join("b.txt"), unstage_root.join("c.txt"))
-            .expect("rename b to c in workdir");
-
-        for path in action_paths_for_file(&unstage_root, "b.txt", GitStatusLayer::Index) {
-            run_git_command(&unstage_root, &["restore", "--staged", "--", &path])
-                .await
-                .expect("unstage index rename path");
-        }
-        let unstage_index = open_repository_at_root(&unstage_root)
-            .expect("open unstaged repo")
-            .index()
-            .expect("open unstaged index");
-        assert!(unstage_index.get_path(Path::new("a.txt"), 0).is_some());
-        assert!(unstage_index.get_path(Path::new("b.txt"), 0).is_none());
-        assert!(unstage_index.get_path(Path::new("c.txt"), 0).is_none());
-    }
-
-    fn index_blob_text(repo_root: &Path, path: &str) -> String {
-        let repo = open_repository_at_root(repo_root).expect("open repo");
-        let index = repo.index().expect("open index");
-        let entry = index
-            .get_path(Path::new(path), 0)
-            .unwrap_or_else(|| panic!("missing index path {path}"));
-        let blob = repo.find_blob(entry.id).expect("find index blob");
-        String::from_utf8(blob.content().to_vec()).expect("index blob utf8")
-    }
-
-    #[tokio::test]
-    async fn unstaged_restore_keeps_staged_hunks_on_mixed_file() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("mixed.txt"), "head\n").expect("write head content");
-        commit_all_with_message(&root, "init mixed fixture").await;
-
-        fs::write(root.join("mixed.txt"), "staged\n").expect("write staged content");
-        run_git_command(&root, &["add", "--", "mixed.txt"])
-            .await
-            .expect("stage mixed file");
-        fs::write(root.join("mixed.txt"), "unstaged\n").expect("write unstaged content");
-
-        for path in action_paths_for_file(&root, "mixed.txt", GitStatusLayer::Workdir) {
-            run_git_command(&root, &["restore", "--worktree", "--", &path])
-                .await
-                .expect("restore unstaged worktree");
-        }
-
-        assert_eq!(
-            fs::read_to_string(root.join("mixed.txt")).expect("read worktree"),
-            "staged\n"
-        );
-        assert_eq!(index_blob_text(&root, "mixed.txt"), "staged\n");
-
-        let repo = open_repository_at_root(&root).expect("open mixed repo");
-        let mut status_options = StatusOptions::new();
-        status_options
-            .include_untracked(true)
-            .include_ignored(false);
-        let statuses = repo
-            .statuses(Some(&mut status_options))
-            .expect("collect mixed statuses");
-        let entry = statuses
-            .iter()
-            .find(|entry| entry.path() == Some("mixed.txt"))
-            .expect("mixed.txt status");
-        assert!(
-            entry.status().contains(Status::INDEX_MODIFIED),
-            "staged hunks must remain after unstaged discard"
-        );
-        assert!(
-            !entry
-                .status()
-                .intersects(Status::WT_MODIFIED | Status::WT_DELETED | Status::WT_NEW),
-            "working tree must match the index after unstaged discard"
-        );
-    }
-
-    #[tokio::test]
-    async fn unstaged_restore_leaves_untracked_for_clean_fallback() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("tracked.txt"), "tracked\n").expect("write tracked content");
-        commit_all_with_message(&root, "init untracked fixture").await;
-        fs::write(root.join("scratch.txt"), "untracked\n").expect("write untracked file");
-
-        let restore_result =
-            run_git_command(&root, &["restore", "--worktree", "--", "scratch.txt"]).await;
-        assert!(
-            restore_result.is_err(),
-            "untracked paths should fail restore so discard can fall back to clean"
-        );
-        run_git_command(&root, &["clean", "-f", "--", "scratch.txt"])
-            .await
-            .expect("clean untracked path");
-        assert!(!root.join("scratch.txt").exists());
-        assert_eq!(
-            fs::read_to_string(root.join("tracked.txt")).expect("read tracked file"),
-            "tracked\n"
-        );
-    }
-
-    #[tokio::test]
-    async fn revert_all_still_discards_staged_and_unstaged() {
-        let (root, _repo) = create_temp_repo();
-        fs::write(root.join("keep.txt"), "head\n").expect("write head content");
-        commit_all_with_message(&root, "init revert-all fixture").await;
-        fs::write(root.join("keep.txt"), "staged\n").expect("write staged content");
-        run_git_command(&root, &["add", "--", "keep.txt"])
-            .await
-            .expect("stage keep file");
-        fs::write(root.join("keep.txt"), "unstaged\n").expect("write unstaged content");
-        fs::write(root.join("scratch.txt"), "untracked\n").expect("write untracked file");
-
-        run_git_command(&root, &["restore", "--staged", "--worktree", "--", "."])
-            .await
-            .expect("revert all restore");
-        run_git_command(&root, &["clean", "-f", "-d"])
-            .await
-            .expect("revert all clean");
-
-        assert_eq!(
-            fs::read_to_string(root.join("keep.txt")).expect("read after revert all"),
-            "head\n"
-        );
-        assert!(!root.join("scratch.txt").exists());
-        assert_worktree_clean(&root);
-    }
-
-    #[test]
-    fn open_repository_at_root_does_not_search_parent_directories() {
-        let (root, _repo) = create_temp_repo();
-        let nested = root.join("nested").join("non-repo");
-        fs::create_dir_all(&nested).expect("create nested directory");
-
-        let result = open_repository_at_root(&nested);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn paginate_history_commits_respects_offset_and_limit() {
-        let commits = (0..5)
-            .map(|index| GitHistoryCommit {
-                sha: format!("sha-{index}"),
-                short_sha: format!("s{index}"),
-                summary: format!("commit-{index}"),
-                message: format!("message-{index}"),
-                author: "tester".to_string(),
-                author_email: "tester@example.com".to_string(),
-                timestamp: 100 + index as i64,
-                parents: Vec::new(),
-                refs: Vec::new(),
-                file_path: None,
-            })
-            .collect::<Vec<_>>();
-        let (page, total, has_more) = paginate_history_commits(commits, 2, 2);
-        assert_eq!(total, 5);
-        assert_eq!(page.len(), 2);
-        assert_eq!(page[0].sha, "sha-2");
-        assert_eq!(page[1].sha, "sha-3");
-        assert!(has_more);
-    }
-
-    #[test]
-    fn validate_local_branch_name_allows_slash_and_rejects_invalid() {
-        assert_eq!(
-            validate_local_branch_name("feature/git-log").expect("valid branch"),
-            "feature/git-log".to_string()
-        );
-        assert!(validate_local_branch_name("feature..broken").is_err());
-    }
-
-    #[test]
-    fn detect_used_by_worktree_delete_error() {
-        let message =
-            "error: cannot delete branch 'feature/test' used by worktree at '/tmp/worktree'";
-        assert!(is_branch_used_by_worktree_error(message));
-        assert_eq!(
-            extract_worktree_path_from_delete_error(message).as_deref(),
-            Some("/tmp/worktree")
-        );
-    }
-
-    #[test]
-    fn build_actionable_used_by_worktree_error_with_path() {
-        let message =
-            "error: cannot delete branch 'feature/test' used by worktree at '/tmp/worktree'";
-        let friendly = build_delete_branch_worktree_error("feature/test", message);
-        assert!(friendly.contains("Switch that worktree to another branch"));
-        assert!(friendly.contains("/tmp/worktree"));
-    }
-
-    #[test]
-    fn token_isolated_command_display_includes_env_unset_prefix() {
-        let rendered = build_token_isolated_command_display(
-            "git",
-            &[
-                "push".to_string(),
-                "-u".to_string(),
-                "origin".to_string(),
-                "HEAD:feature/a".to_string(),
-            ],
-        );
-        assert!(rendered.starts_with("env -u GH_TOKEN -u GITHUB_TOKEN"));
-        assert!(rendered.contains("git push -u origin HEAD:feature/a"));
-    }
-
-    #[test]
-    fn detect_http2_transport_error_signature() {
-        let message = "error: RPC failed; curl 16 Error in the HTTP2 framing layer";
-        assert!(is_http2_transport_error(message));
-    }
-
-    #[test]
-    fn parse_pr_number_from_url_works() {
-        assert_eq!(
-            extract_pr_number_from_url("https://github.com/a/b/pull/123"),
-            Some(123)
-        );
-        assert_eq!(
-            extract_pr_number_from_url("https://github.com/a/b/pull/456/files"),
-            Some(456)
-        );
-        assert_eq!(
-            extract_pr_number_from_url("https://github.com/a/b/issues/1"),
-            None
-        );
-    }
-
-    #[test]
-    fn range_gate_passes_normal_threshold() {
-        let range_fingerprint = "base-revision...head-revision";
-        let paths = (0..range_gate::PR_RANGE_MAX_CHANGED_FILES)
-            .map(|index| format!("src/file-{index}.ts"))
-            .collect::<Vec<_>>();
-        let decision = evaluate_pr_range_gate(&paths, false, None, range_fingerprint);
-        assert!(matches!(
-            decision,
-            PrRangeGateDecision::Pass { changed_files }
-                if changed_files == range_gate::PR_RANGE_MAX_CHANGED_FILES
-        ));
-    }
-
-    #[test]
-    fn range_gate_requires_confirmation_above_normal_threshold() {
-        let range_fingerprint = "base-revision...head-revision";
-        for changed_files in [
-            range_gate::PR_RANGE_MAX_CHANGED_FILES + 1,
-            range_gate::PR_RANGE_COMPLETE_DIFF_MAX_FILES,
-        ] {
-            let paths = (0..changed_files)
-                .map(|index| format!("src/file-{index}.ts"))
-                .collect::<Vec<_>>();
-            let decision = evaluate_pr_range_gate(&paths, false, None, range_fingerprint);
-            assert!(matches!(
-                decision,
-                PrRangeGateDecision::ConfirmationRequired {
-                    range_gate: GitPrRangeGate {
-                        severity: GitPrRangeGateSeverity::Large,
-                        range_fingerprint: fingerprint,
-                        ..
-                    },
-                    ..
-                } if fingerprint == range_fingerprint
-            ));
-        }
-    }
-
-    #[test]
-    fn range_gate_warns_when_github_diff_will_be_incomplete() {
-        let range_fingerprint = "base-revision...head-revision";
-        let paths = (0..(range_gate::PR_RANGE_COMPLETE_DIFF_MAX_FILES + 1))
-            .map(|index| format!("src/file-{index}.ts"))
-            .collect::<Vec<_>>();
-        let decision = evaluate_pr_range_gate(&paths, false, None, range_fingerprint);
-        assert!(matches!(
-            decision,
-            PrRangeGateDecision::ConfirmationRequired {
-                range_gate: GitPrRangeGate {
-                    severity: GitPrRangeGateSeverity::DiffIncomplete,
-                    ..
-                },
-                ..
-            }
-        ));
-        assert!(matches!(
-            evaluate_pr_range_gate(
-                &paths,
-                true,
-                Some(range_fingerprint),
-                range_fingerprint
-            ),
-            PrRangeGateDecision::Pass { changed_files }
-                if changed_files == range_gate::PR_RANGE_COMPLETE_DIFF_MAX_FILES + 1
-        ));
-        assert!(matches!(
-            evaluate_pr_range_gate(
-                &paths,
-                true,
-                Some("stale-base...stale-head"),
-                range_fingerprint
-            ),
-            PrRangeGateDecision::ConfirmationRequired { .. }
-        ));
-    }
-
-    #[test]
-    fn range_gate_override_cannot_bypass_structural_anomalies() {
-        let range_fingerprint = "base-revision...head-revision";
-        assert!(matches!(
-            evaluate_pr_range_gate(
-                &[],
-                true,
-                Some(range_fingerprint),
-                range_fingerprint
-            ),
-            PrRangeGateDecision::Blocked { category, .. } if category == "range-empty"
-        ));
-
-        let mut paths = (0..range_gate::PR_RANGE_SUSPICIOUS_THRESHOLD)
-            .map(|index| format!("src/file-{index}.ts"))
-            .collect::<Vec<_>>();
-        paths.push("README.md".to_string());
-        let decision =
-            evaluate_pr_range_gate(&paths, true, Some(range_fingerprint), range_fingerprint);
-        assert!(matches!(
-            decision,
-            PrRangeGateDecision::Blocked { category, .. } if category == "range-suspicious"
-        ));
-    }
-
-    #[test]
-    fn range_fingerprint_requires_exactly_base_and_head_revisions() {
-        assert_eq!(
-            parse_pr_range_fingerprint("base-revision\nhead-revision\n"),
-            Some("base-revision...head-revision".to_string())
-        );
-        assert_eq!(parse_pr_range_fingerprint("base-only\n"), None);
-        assert_eq!(parse_pr_range_fingerprint("base\nhead\nunexpected\n"), None);
-    }
-}
+mod tests;
